requirements.txt
requests
kafka-python
pymongo
cassandra-driver
python-dotenv
